Move the long crud tests module into a crud submodule to separate tests. In future can break tests by logic more.
